% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biroot.R
\name{biroot}
\alias{biroot}
\title{Find points on boundary of discrete classification}
\value{
A dataframe with a row for each point with columns giving, location,
how deep in the grid search they were, a unique id for each square,
and finally the result of the discrete function at that point.
}
\description{
Find points on boundary of discrete classification
}
\section{Arguments}{
 \code{biroot()} understands the following arguments:
\itemize{
\item sq - corners of the region to be searched
\item f - function for discrete classification over region
\item max_depth - how many times should the grid be split in four
\item min_depth - how many times should the grid split before becoming adaptive
\item ... - additional arguments for f
}
}

\examples{
# using it to plot boundaries of qda

set.seed(17)
1+1
}
