% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biroot_lines.R
\name{biroot_lines}
\alias{biroot_lines}
\title{Return lines approximating the solution set of the function given
within the region given.}
\usage{
biroot_lines(
  f,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  max_depth = 10,
  min_depth = 2,
  class = "continuous",
  ...
)
}
\value{
A dataframe with a row for each point with columns giving, location,
how deep in the grid search they were, a unique id for pair of points.
}
\description{
Return lines approximating the solution set of the function given
within the region given.
}
\section{Arguments}{
 \code{biroot_lines()} understands the following arguments:
\itemize{
\item f - function with zero set within region
\item xlim - boundaries of x to be searched
\item ylim - boundaries of y to be searched
\item max_depth - how many times should the grid be split in four
\item min_depth - how many times should the grid split before becoming adaptive
\item ... - additional arguments for f
}
}

\examples{

ftest <- function(x) with(x, x^2+y^2-1)
solution_lines <- biroot_lines(ftest, xlim = c(-2,2), ylim = c(-2,2))
solution_lines |> str()

solution_lines |> 
  ggplot(aes(x, y, group = id)) +
    geom_line() +
    coord_equal()
  
}
