% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{classify_multi}
\alias{classify_multi}
\title{Plot multiple different methods for \code{geom_classify} at once utilizing patchwork}
\arguments{
\item{methods}{a list of methods to be plotted.}

\item{titles}{a vector of strings that will be used for plot titles in the same
order that methods are selected.}

\item{boundary}{if set to true \code{geom_classify_boundary} will be used instead.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{baseplot}{the core ggplot object added to by each method}
}
\description{
Perform model estimation and compute the estimated class of every point
on a grid. \code{geom_classify} draws filled regions.
}
