% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggclassify-package.R
\docType{package}
\name{ggclassify}
\alias{ggclassify-package}
\alias{ggclassify}
\alias{package-ggclassify}
\title{ggclassify: Stats and Geoms for Classification with ggplot2}
\description{
Description of package
}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/nathaniel-morgan/ggclassify/}
}
}
\author{
\strong{Maintainer}: Nathaniel Morgan \email{nathaniel.morgan1999@gmail.com} (\href{https://orcid.org/0009-0008-1397-8522}{ORCID})

Authors:
\itemize{
  \item David Kahle \email{david@kahle.io} (\href{https://orcid.org/0000-0002-9999-1558}{ORCID})
  \item Michael Gallaugher \email{michael_gallaugher@baylor.edu} (\href{https://orcid.org/0000-0002-5487-8965}{ORCID})
}

}
