% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{pca}
\alias{pca}
\title{Compute n principle components for a data frame and append them as
new columns. Does not utilize the class variable you are interested
in plotting for.}
\arguments{
\item{x}{data frame to compute principal components for.}

\item{class}{The name of the class variable you don't want included in the PC's.}

\item{scale}{Whether to center and scale the data before Computing PC's.}

\item{exclude}{character vector of any other columns to excluded from PC's.}

\item{n}{The number of PC's to compute and append.}
}
\description{
Compute n principle components for a data frame and append them as
new columns. Does not utilize the class variable you are interested
in plotting for.
}
