% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadroot_discrete.R
\name{quadroot_discrete}
\alias{quadroot_discrete}
\title{Find points on boundary of discrete classification}
\description{
Find points on boundary of discrete classification
}
\section{Arguments}{
 \code{quadroot_discrete()} understands the following arguments:
\itemize{
\item x - The region to be searched
}
}

\examples{
# Basic test

set.seed(17)
1+1

}
