% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biroot-package.R, R/biroot.R
\docType{package}
\name{biroot}
\alias{biroot-package}
\alias{biroot}
\alias{package-biroot}
\title{biroot: quadtree inspired method for approximating solution curves.}
\usage{
biroot(
  f,
  xlim = c(-1, 1),
  ylim = c(-1, 1),
  max_depth = 10,
  min_depth = 2,
  class = "continuous",
  ...
)
}
\value{
A dataframe with a row for each point with columns giving, location,
how deep in the grid search they were, a unique id for each square,
and finally the result of the discrete function at that point.
}
\description{
Description of package
}
\section{Arguments}{
 \code{biroot()} understands the following arguments:
\itemize{
\item f - function for discrete classification over region
\item xlim - boundaries of x to be searched
\item ylim - boundaries of y to be searched
\item max_depth - how many times should the grid be split in four
\item min_depth - how many times should the grid split before becoming adaptive
\item ... - additional arguments for f
}
}

\examples{

ftest <- function(x) with(x, x^2-y-2)
biroot(ftest, xlim = c(-5,5), ylim = c(-5,5))
biroot(ftest, xlim = c(-5,5), ylim = c(-5,5)) |> str()

biroot(ftest, xlim = c(-5,5), ylim = c(-5,5)) |> 
  ggplot(aes(x, y)) +
    geom_point(aes(fill = value), shape = 21) +
    coord_equal()
  
ftest <- function(v) with(v, x^2 + y^2 - 1)

(df <- biroot(ftest, xlim = c(-2,2), ylim = c(-2,2), max_depth = 9) )

df |> 
  ggplot(aes(x, y)) +
    geom_point(aes(fill = value), shape = 21) +
    geom_contour(
      aes(z = z),
      color = "red",
      breaks = 0, 
      data = expand.grid(
        "x" = seq(-5, 5, length.out = 101), 
        "y" = seq(-5, 5, length.out = 101)
      ) |> 
        transform("z" = x^2 + y^2 - 1)
    ) +  
    coord_equal()

}
\seealso{
Useful links:
\itemize{
\item \url{https://github.com/nathaniel-morgan/biroot/}
}
}
\author{
\strong{Maintainer}: Nathaniel Morgan \email{nathaniel.morgan1999@gmail.com} (\href{https://orcid.org/0009-0008-1397-8522}{ORCID})

Authors:
\itemize{
  \item David Kahle \email{david@kahle.io} (\href{https://orcid.org/0000-0002-9999-1558}{ORCID})
  \item Michael Gallaugher \email{michael_gallaugher@baylor.edu} (\href{https://orcid.org/0000-0002-5487-8965}{ORCID})
}

}
